Fix a bug in the resolver which always resolves to earliest scope
The resolver overwrites the distances of all names and wrongfully always points
to the earliest scope. This results in variables always resolving to the global
or the first scope and all following scopes are ignored.

This commit extends the scope test and makes it easier to track the desired behaviour.